interviewace/backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ── Database ───────────────────────────────────────────────────
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
# psycopg2-binary   # uncomment when using PostgreSQL

# ── Auth ───────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── AI / NLP ───────────────────────────────────────────────────
anthropic>=0.28.0
requests>=2.32.0

# ── File parsing ───────────────────────────────────────────────
PyPDF2>=3.0.1
python-docx>=1.1.2

# ── Utilities ──────────────────────────────────────────────────
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.5.0
